Refactor the interactive mode, add shell.py
