+++ docs/rivescript.rst
@@ -12,6 +12,14 @@ rivescript.rivescript module
     :undoc-members:
     :show-inheritance:
 
+rivescript.brain module
+-----------------------
+
+.. automodule:: rivescript.brain
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 rivescript.exceptions module
 ----------------------------
 
@@ -20,11 +28,58 @@ rivescript.exceptions module
     :undoc-members:
     :show-inheritance:
 
+rivescript.inheritance module
+-----------------------------
+
+.. automodule:: rivescript.inheritance
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+rivescript.interactive module
+-----------------------------
+
+.. automodule:: rivescript.interactive
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+rivescript.parser module
+------------------------
+
+.. automodule:: rivescript.parser
+    :members:
+    :undoc-members:
+    :show-inheritance:
 
 rivescript.python module
 ------------------------
 
 .. automodule:: rivescript.python
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+rivescript.regexp module
+------------------------
+
+.. automodule:: rivescript.regexp
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+rivescript.sorting module
+-------------------------
+
+.. automodule:: rivescript.sorting
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+rivescript.utils module
+-----------------------
+
+.. automodule:: rivescript.utils
     :members:
     :undoc-members:
     :show-inheritance:
